+++ CHANGELOG.rst
@@ -1,9 +1,21 @@
 Changelog
 ---------
 
+version 0.5
+
+* **backwards-incompatible change:** each client now conforms to a
+  standard interface, and can be instantiated with the get_client
+  factory method
+
+version 0.4
+
+* each client now logs HTTP GET and POST operations
+
 version 0.3.1
 
-* Prowl.Client.retrieve_token and Prowl.Client.retrieve_apikey now raise a pushnotify.exceptions.ProviderKeyError if the specified Provider Key is invalid
+* Prowl.Client.retrieve_token and Prowl.Client.retrieve_apikey now
+  raise a pushnotify.exceptions.ProviderKeyError if the specified
+  Provider Key is invalid
 
 version 0.3
 
@@ -11,7 +23,8 @@ version 0.3
 
 version 0.2.1
 
-* fixed an issue where developer keys were not being sent with Notify My Android verifications
+* fixed an issue where developer keys were not being sent with Notify My
+  Android verifications
 
 version 0.2